Refactor lobby player methods

+++ lib/croc/games/monopoly/lobby/player.ex
@@ -7,6 +7,7 @@ defmodule Croc.Games.Monopoly.Lobby.Player do
     :lobby_id
   ]
 
+  @derive Jason.Encoder
   use Memento.Table,
     attributes: [
       :id,
@@ -20,7 +21,7 @@ defmodule Croc.Games.Monopoly.Lobby.Player do
   def new(player_id, lobby_id), do: %__MODULE__{player_id: player_id, lobby_id: lobby_id}
 
   def create(player_id, lobby_id) do
-    Memento.transaction(fn ->
+    {:ok, player} = Memento.transaction(fn ->
       Memento.Query.write(%__MODULE__{
         player_id: player_id,
         lobby_id: lobby_id
@@ -28,21 +29,25 @@ defmodule Croc.Games.Monopoly.Lobby.Player do
     end)
   end
 
-  def delete(player_id, lobby_id) do
-    {:ok} =
-      Memento.transaction(fn ->
-        lobby_player =
-          Memento.Query.select(
-            %__MODULE__{
-              player_id: player_id,
-              lobby_id: lobby_id
-            },
-            lobby_id
-          )
-          |> Enum.find(fn p -> p.player_id == player_id end)
+  def delete_by_id(id) do
+    :ok = Memento.transaction(fn ->
+      Memento.Query.delete(__MODULE__, id)
+    end)
+  end
 
+  def delete(player_id, lobby_id) do
+    Memento.transaction(fn ->
+      Logger.error("Select here")
+      lobby_player =
+        Memento.Query.select(__MODULE__, {:==, :lobby_id, lobby_id})
+        |> Enum.find(fn p -> p.player_id == player_id end)
+
+      unless lobby_player == nil do
         Memento.Query.delete(__MODULE__, lobby_player.id)
-      end)
+      else
+        :not_found
+      end
+    end)
   end
 
   def already_in_lobby?(player_id) do
